+++ features/administrators/adding_new_administrator_account_with_administration_role.feature
@@ -8,8 +8,8 @@ Feature: Adding a new Administrator account with administration role
         Given the store operates on a single channel in "United States"
         And there is already an Administration role "No sections access" in the system
 
-    @ui
+    @domain
     Scenario: Adding a new administrator account with administration role in the system
         When I add a new administrator "good.father@abc.com" named "Good" with "Father" last name and "No sections access" administration role
-        Then administrator account "good.father@abc.com" should have first name "Good" and last name "Father"
+        Then the administrator account "good.father@abc.com" should have "Good" first name and "Father" last name
         And this administrator should have administration role "No sections access"
