+++ tests/Behat/Context/UI/ManagingAdministratorsContext.php
@@ -5,14 +5,8 @@ declare(strict_types=1);
 namespace Tests\Sylius\RbacPlugin\Behat\Context\UI;
 
 use Behat\Behat\Context\Context;
-use Doctrine\Common\Persistence\ObjectManager;
 use Sylius\Behat\Page\Admin\Administrator\UpdatePageInterface;
-use Sylius\Behat\Service\SharedStorageInterface;
-use Sylius\Bundle\CoreBundle\Fixture\Factory\ExampleFactoryInterface;
-use Sylius\Component\Core\Model\AdminUserInterface;
 use Sylius\Component\User\Repository\UserRepositoryInterface;
-use Sylius\RbacPlugin\Entity\AdministrationRoleAwareInterface;
-use Sylius\RbacPlugin\Entity\AdministrationRoleInterface;
 use Tests\Sylius\RbacPlugin\Behat\Element\AdministrationRolesElementInterface;
 use Tests\Sylius\RbacPlugin\Behat\Page\Ui\AdminUserIndexPageInterface;
 use Webmozart\Assert\Assert;
@@ -31,54 +25,16 @@ final class ManagingAdministratorsContext implements Context
     /** @var UserRepositoryInterface */
     private $adminUserRepository;
 
-    /** @var ObjectManager */
-    private $administratorManager;
-
-    /** @var ExampleFactoryInterface */
-    private $adminUserExampleFactory;
-
-    /** @var SharedStorageInterface */
-    private $sharedStorage;
-
     public function __construct(
         AdministrationRolesElementInterface $administrationRolesElement,
         AdminUserIndexPageInterface $indexPage,
         UpdatePageInterface $updatePage,
-        UserRepositoryInterface $adminUserRepository,
-        ObjectManager $administratorManager,
-        ExampleFactoryInterface $adminUserExampleFactory,
-        SharedStorageInterface $sharedStorage
+        UserRepositoryInterface $adminUserRepository
     ) {
         $this->administrationRolesElement = $administrationRolesElement;
         $this->indexPage = $indexPage;
         $this->updatePage = $updatePage;
         $this->adminUserRepository = $adminUserRepository;
-        $this->administratorManager = $administratorManager;
-        $this->adminUserExampleFactory = $adminUserExampleFactory;
-        $this->sharedStorage = $sharedStorage;
-    }
-
-    /**
-     * @When I add a new administrator :email named :firstName with :lastname last name and :administrationRole administration role
-     */
-    public function iAddANewAdministratorNamedWithLastNameAndAdministrationRole(
-        string $email,
-        string $firstName,
-        string $lastname,
-        AdministrationRoleInterface $administrationRole
-    ): void {
-        $adminUser = $this->adminUserExampleFactory->create([
-            'email' => $email,
-            'username' => $email,
-            'password' => $email,
-            'enabled' => true,
-            'first_name' => $firstName,
-            'last_name' => $lastname,
-            'administration_role' => $administrationRole
-        ]);
-
-        $this->adminUserRepository->add($adminUser);
-        $this->sharedStorage->set('administrator', $adminUser);
     }
 
     /**
@@ -122,21 +78,5 @@ final class ManagingAdministratorsContext implements Context
     public function administratorShouldHaveNoRoleAssigned(string $email): void
     {
         Assert::same('', $this->indexPage->getAdministrationRoleOfAdminWithEmail($email));
-    }
-
-    /**
-     * @Then administrator account :email should have first name :firstName and last name :lastname
-     */
-    public function administratorAccountShouldHaveFirstNameAndLastName(string $email, string $firstName, string $lastname): void
-    {
-        /** @var AdminUserInterface $user */
-        $user = $this->adminUserRepository->findOneBy(['email' => $email]);
-        Assert::true(
-            $user->getEmail() === $email &&
-            $user->getFirstName() === $firstName &&
-            $user->getLastName() === $lastname &&
-            $user->getUsername() === $email &&
-            $user->isEnabled() === true
-        );
     }
 }
