+++ tests/Behat/Resources/services/contexts.xml
@@ -29,7 +29,10 @@
             <argument type="service" id="Tests\Sylius\RbacPlugin\Behat\Page\Ui\AdminUserIndexPage" />
             <argument type="service" id="sylius.behat.page.admin.administrator.update" />
             <argument type="service" id="sylius.repository.admin_user" />
-            <argument type="service" id="sylius.manager.admin_user" />
+        </service>
+
+        <service id="Tests\Sylius\RbacPlugin\Behat\Context\Domain\ManagingAdministratorsContext" public="true">
+            <argument type="service" id="sylius.repository.admin_user" />
             <argument type="service" id="sylius.fixture.example_factory.admin_user" />
             <argument type="service" id="sylius.behat.shared_storage" />
         </service>
